solving the endian problem

+++ ping.h
@@ -7,6 +7,7 @@
 #include<sys/socket.h>
 #include "string.h"
 #include "errno.h"
+#include <netinet/in.h>
 
 //colors to use
 
@@ -31,7 +32,7 @@ typedef struct h_icmp{
 icmp *create_icmp_packet(uint8 type,uint8 code,uint8 *data,uint16 size);
 void error(const int8* msg);
 uint16 checksum(icmp *,uint16 size);
-uint16 _checksum(uint16 *data,size_t n);
+uint16 _checksum(uint8 *data,size_t n);
 void print_icmp_packet(icmp *packet,uint16);
 
 #endif
